Rename DAG util to utils for consistency
