+++ airflow_home/dags/webhooks_dags.py
@@ -6,7 +6,7 @@ from fn.func import F
 import stringcase as case
 import pymongo
 import os
-from util.docker import create_docker_operator, create_linked_docker_operator
+from utils.docker import create_docker_operator, create_linked_docker_operator
 
 now = datetime.utcnow() - timedelta(hours=1)
 start_date = datetime(now.year, now.month, now.day, now.hour)
